# openapi-v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VolumeMeta": {
        "properties": {
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "dimensions": {
            "items": {
              "type": "integer"
            },
            "title": "Dimensions",
            "type": "array"
          },
          "direction": {
            "items": {
              "type": "number"
            },
            "title": "Direction",
            "type": "array"
          },
          "hu_max": {
            "title": "Hu Max",
            "type": "integer"
          },
          "hu_min": {
            "title": "Hu Min",
            "type": "integer"
          },
          "levels": {
            "items": {
              "type": "integer"
            },
            "title": "Levels",
            "type": "array"
          },
          "origin": {
            "items": {
              "type": "number"
            },
            "title": "Origin",
            "type": "array"
          },
          "spacing_mm": {
            "items": {
              "type": "number"
            },
            "title": "Spacing Mm",
            "type": "array"
          }
        },
        "required": [
          "dataset_id",
          "dimensions",
          "spacing_mm",
          "origin",
          "direction",
          "hu_min",
          "hu_max",
          "levels"
        ],
        "title": "VolumeMeta",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Axioma Sidecar",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/datasets/{dataset_id}/meta": {
      "get": {
        "operationId": "get_volume_meta_api_v1_datasets__dataset_id__meta_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolumeMeta"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Volume Meta"
      }
    },
    "/api/v1/datasets/{dataset_id}/slice": {
      "get": {
        "operationId": "get_slice_api_v1_datasets__dataset_id__slice_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "axis",
            "required": true,
            "schema": {
              "title": "Axis",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "index",
            "required": true,
            "schema": {
              "title": "Index",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "lod",
            "required": false,
            "schema": {
              "default": "native",
              "title": "Lod",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Slice"
      }
    },
...
    "/api/v1/datasets/{dataset_id}/volume": {
      "get": {
        "operationId": "get_volume_api_v1_datasets__dataset_id__volume_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Level",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Range",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Range"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Volume"
      }
    },
...
  }
}